+++ node/src/components/small_network/error.rs
@@ -1,13 +1,12 @@
 use std::{error, io, net::SocketAddr, result, sync::Arc};
 
-use casper_types::{ProtocolVersion, SecretKey};
+use casper_types::{crypto, ProtocolVersion, SecretKey};
 use datasize::DataSize;
 use openssl::{error::ErrorStack, ssl};
 use serde::Serialize;
 use thiserror::Error;
 
 use crate::{
-    crypto,
     tls::ValidationError,
     utils::{LoadError, Loadable, ResolveAddressError},
 };